+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.3] - 2025-07-30
+
+### Changed
+
+- `RowMajor` and `ColMajor` are back to variantless enums
+- The struct `LayoutKind` was added, which provides dynamic dispatch for layouts
+
 ## [0.5.2] - 2025-07-30
+
+_**⚠️ Yanked**: Replaced by `0.5.3` to avoid a breaking change._
 
 ### Changed